pyproject.toml

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "hands-on-gcp-closer-academy"
version = "0.2.0"
description = "Hands On Lab using GCP"
authors = ["João Monteiro, Ricardo Miranda"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-storage = "^2.10.0"
google-cloud-bigquery = "^3.11.4"
google-cloud-pubsub = "^2.18.4"
google-cloud-firestore = "^2.11.1"

[tool.poetry.group.cloudfunctions.dependencies]
functions-framework = "^3.0.0"
cloudevents = "^1.9.0"
deprecation = "2.1.0"
Flask = "^2.2.5"
gunicorn = "20.1.0"
importlib-metadata = "6.7.0"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.3"
packaging = "23.1"
typing_extensions = "4.7.1"
Werkzeug = "2.2.3"
zipp = "3.15.0"

[tool.poetry.group.model_train.dependencies]
scikit-learn = "^1.3.0"
pandas = "^2.0.0"
db-dtypes = "^1.1.1"

[tool.poetry.group.predictions.dependencies]
joblib = "^1.3.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
mypy = "^1.5.1"
black = "^24.4.2"
deepdiff = "^6.4.1"
pytest = "^7.4.2"
pre-commit = "^3.4.0"
pytest-env = "^1.0.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "1.5.3"
mkdocstrings-python = "*"
mkdocstrings = "*"
mkdocs-material = "*"

[tool.pytest.ini_options]
env = [
    "_CI_TESTING=yes",
]

[tool.ruff]
fix = true
line-length = 150
ignore = ["E722"]
select = ["F", "E", "W", "I"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
select = ["D"]
ignore = ["D206"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
pretty = true
namespace_packages = true
files = "functions/**/*.py"
ignore_missing_imports = false
strict_optional = false
warn_unused_configs = true
disallow_untyped_defs = true
disallow_untyped_calls = true
check_untyped_defs = true
follow_imports = "normal"

[tool.black]
line-length = 150
target-version = ['py311']
include = '\.pyi?$'